+++ src/main/java/bing/system/service/SysMenuServiceImpl.java
@@ -39,8 +39,8 @@ public class SysMenuServiceImpl implements SysMenuService {
 
 	@Override
 	public GenericPage<SysMenuVO> listByPage(SysMenuCondition condition) {
-		Long pageNumber = condition.getPageNumber();
-		PageHelper.startPage(pageNumber.intValue(), condition.getPageSize().intValue());
+		int pageNumber = condition.getPageNumber();
+		PageHelper.startPage(pageNumber, condition.getPageSize());
 		List<SysMenuVO> list = sysMenuDao.listByCondition(condition);
 		PageInfo<SysMenuVO> pageInfo = new PageInfo<>(list);
 		return new GenericPage<>(pageNumber, pageInfo.getTotal(), list);
@@ -83,7 +83,7 @@ public class SysMenuServiceImpl implements SysMenuService {
 	}
 
 	@Override
-	@Cacheable(cacheNames = {EhCacheNames.MENU_TREE_CACHE})
+	@Cacheable(cacheNames = { EhCacheNames.MENU_TREE_CACHE })
 	public List<MenuTreeNode> getMenuTree() {
 		List<SysMenuVO> topMenus = sysMenuDao.listByParentId(GlobalConstants.TOP_PARENT_ID);
 		List<SysMenuVO> menus = sysMenuDao.listAll();
@@ -100,7 +100,8 @@ public class SysMenuServiceImpl implements SysMenuService {
 		List<MenuTreeNode> treeNodes = convertMenu(topMenus);
 		treeNodes.forEach(menu -> menu.setIconSkin(GlobalConstants.ICON_CLS_ROOT));
 		List<MenuTreeNode> allTreeNodes = convertMenu(menus);
-		List<MenuTreeNode> treeNodesExclude = allTreeNodes.stream().filter(treeNode -> !StringUtils.equals(treeNode.getId(), Objects.toString(id))).collect(Collectors.toList());
+		List<MenuTreeNode> treeNodesExclude = allTreeNodes.stream().filter(treeNode -> !StringUtils.equals(treeNode.getId(), Objects.toString(id)))
+				.collect(Collectors.toList());
 		MenuTreeNode.buildMenuTree(treeNodes, treeNodesExclude);
 
 		return treeNodes;
@@ -112,7 +113,8 @@ public class SysMenuServiceImpl implements SysMenuService {
 		List<SysMenuVO> menus = sysMenuDao.listAll();
 		List<Integer> ownResourceIds = sysResourceDao.listResourceIdByUserId(userId);
 		// 迭代全部菜单，未绑定资源的忽略，绑定资源的需要与用户具备的资源比较，不包含则删除菜单
-		List<SysMenuVO> ownMenus = menus.stream().filter(menu -> (menu.getResourceId() == null) || (ownResourceIds.contains(menu.getResourceId()))).collect(Collectors.toList());
+		List<SysMenuVO> ownMenus = menus.stream().filter(menu -> (menu.getResourceId() == null) || (ownResourceIds.contains(menu.getResourceId())))
+				.collect(Collectors.toList());
 		List<MenuTreeNode> treeNodes = convertMenuWithUrl(topMenus);
 		MenuTreeNode.buildMenuTree(treeNodes, convertMenuWithUrl(ownMenus));
 		// 移除顶级空菜单，即没子菜单的菜单
